services/admin/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.0
h11==0.14.0
idna==3.8
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic-core==2.23.2
python-decouple==3.8
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.4
typing-extensions==4.12.2
uvicorn==0.30.6
